cleaning up examples
